Changed Contact Page
Input and textarea no longer have blue outline when selecting (only a small thin one) as well as adding text stating * are required and changing the placeholders. Name placeholder was "" now it is "First and Last Name". Message placeholder used to be "..." now it is "What you would like us to know"